overhauled the network tabs, so that following/followers are seperate, and then adjusted the state management so that following/unfollowing people updates things properly
