+++ client/src/Components/Feeds/Posts/LazyPostTab.js
@@ -1,5 +1,5 @@
 import React, { useState, Suspense, lazy, useEffect } from 'react'
-import Loading from '../Loading'
+import Loading from '../../Loading'
 const LazyPostTab = lazy(() => import('./PostTab'))
 
 function component(props) {
