+++ client/src/Components/Feeds/Posts/PostTab.js
@@ -1,7 +1,7 @@
 import { React, useState, useEffect, useContext } from 'react'
 import { Collapse } from 'react-bootstrap'
-import { UserContext } from '../../utils/UserState'
-import { socket } from '../../utils/socket'
+import { UserContext } from '../../../utils/UserState'
+import { socket } from '../../../utils/socket'
 import PostBox from './PostBox'
 
 function PostTab({ post, home, expanded }) {
